完成全部服务接口的编写， 工具类的部分调整， 完成部分服务实现类的编写 调整整体代码格式整洁

+++ src/main/java/me/superning/userpassbook/service/MerchantService.java
@@ -34,6 +34,6 @@ public interface MerchantService {
      * @param ids
      * @return {@link Merchant}
      */
-    List<Merchant> findByIdPutIn(List<Integer> ids);
+    List<Merchant> findByIdPutIn(List<Long> ids);
 
 }
